requirements.txt
# Dependencies for the Discord knowledge-base bot (early 2023 build)
# Pinned to the pre-v1.0 openai SDK that was current at build time.
discord.py>=2.0.0,<3.0.0
openai>=0.27.0,<1.0.0
python-docx>=0.8.11
